+++ script/tool/test/lint_android_command_test.dart
@@ -37,79 +37,6 @@ void main() {
       runner.addCommand(command);
     });
 
-    void writeFakePluginBuildGradle(RepositoryPackage plugin,
-        {bool warningsConfigured = true}) {
-      const String warningConfig = '''
-    lintOptions {
-        checkAllWarnings true
-        warningsAsErrors true
-        disable 'AndroidGradlePluginVersion', 'InvalidPackage', 'GradleDependency'
-        baseline file("lint-baseline.xml")
-    }
-''';
-      final File gradleFile = plugin
-          .platformDirectory(FlutterPlatform.android)
-          .childFile('build.gradle');
-      gradleFile.createSync(recursive: true);
-      gradleFile.writeAsStringSync('''
-android {
-    compileSdkVersion 33
-
-    defaultConfig {
-        minSdkVersion 16
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-${warningsConfigured ? warningConfig : ''}
-    testOptions {
-        unitTests.returnDefaultValues = true
-        unitTests.all {
-            testLogging {
-               events "passed", "skipped", "failed", "standardOut", "standardError"
-               outputs.upToDateWhen {false}
-               showStandardStreams = true
-            }
-        }
-    }
-}
-
-''');
-    }
-
-    void writeFakeExampleBuildGradle(
-        RepositoryPackage example, String pluginName,
-        {bool warningsConfigured = true}) {
-      final String warningConfig = '''
-gradle.projectsEvaluated {
-    project(":$pluginName") {
-        tasks.withType(JavaCompile) {
-            options.compilerArgs << "-Xlint:all" << "-Werror"
-        }
-    }
-}
-''';
-      example
-          .platformDirectory(FlutterPlatform.android)
-          .childFile('build.gradle')
-          .writeAsStringSync('''
-buildscript {
-    repositories {
-        google()
-        mavenCentral()
-    }
-    dependencies {
-        classpath 'com.android.tools.build:gradle:8.0.1'
-    }
-}
-allprojects {
-    repositories {
-        google()
-        mavenCentral()
-    }
-}
-${warningsConfigured ? warningConfig : ''}
-''');
-    }
-
     test('runs gradle lint', () async {
       final RepositoryPackage plugin =
           createFakePlugin('plugin1', packagesDir, extraFiles: <String>[
@@ -117,8 +44,6 @@ ${warningsConfigured ? warningConfig : ''}
       ], platformSupport: <String, PlatformDetails>{
         platformAndroid: const PlatformDetails(PlatformSupport.inline)
       });
-      writeFakePluginBuildGradle(plugin);
-      writeFakeExampleBuildGradle(plugin.getExamples().first, 'plugin1');
 
       final Directory androidDir =
           plugin.getExamples().first.platformDirectory(FlutterPlatform.android);
@@ -156,10 +81,6 @@ ${warningsConfigured ? warningConfig : ''}
           platformSupport: <String, PlatformDetails>{
             platformAndroid: const PlatformDetails(PlatformSupport.inline)
           });
-      writeFakePluginBuildGradle(plugin);
-      for (final RepositoryPackage example in plugin.getExamples()) {
-        writeFakeExampleBuildGradle(example, 'plugin1');
-      }
 
       final Iterable<Directory> exampleAndroidDirs = plugin.getExamples().map(
           (RepositoryPackage example) =>
@@ -189,11 +110,10 @@ ${warningsConfigured ? warningConfig : ''}
     });
 
     test('fails if gradlew is missing', () async {
-      final RepositoryPackage plugin = createFakePlugin('plugin1', packagesDir,
+      createFakePlugin('plugin1', packagesDir,
           platformSupport: <String, PlatformDetails>{
             platformAndroid: const PlatformDetails(PlatformSupport.inline)
           });
-      writeFakePluginBuildGradle(plugin);
 
       Error? commandError;
       final List<String> output = await runCapturingPrint(
@@ -218,8 +138,6 @@ ${warningsConfigured ? warningConfig : ''}
       ], platformSupport: <String, PlatformDetails>{
         platformAndroid: const PlatformDetails(PlatformSupport.inline)
       });
-      writeFakePluginBuildGradle(plugin);
-      writeFakeExampleBuildGradle(plugin.getExamples().first, 'plugin1');
 
       final String gradlewPath = plugin
           .getExamples()
@@ -242,63 +160,6 @@ ${warningsConfigured ? warningConfig : ''}
           output,
           containsAllInOrder(
             <Matcher>[
-              contains('The following packages had errors:'),
-            ],
-          ));
-    });
-
-    test('fails if gradle-driven lint-warnings-as-errors is missing', () async {
-      final RepositoryPackage plugin =
-          createFakePlugin('plugin1', packagesDir, extraFiles: <String>[
-        'example/android/gradlew',
-      ], platformSupport: <String, PlatformDetails>{
-        platformAndroid: const PlatformDetails(PlatformSupport.inline)
-      });
-      writeFakePluginBuildGradle(plugin, warningsConfigured: false);
-      writeFakeExampleBuildGradle(plugin.getExamples().first, 'plugin1');
-
-      Error? commandError;
-      final List<String> output = await runCapturingPrint(
-          runner, <String>['lint-android'], errorHandler: (Error e) {
-        commandError = e;
-      });
-
-      expect(commandError, isA<ToolExit>());
-      expect(
-          output,
-          containsAllInOrder(
-            <Matcher>[
-              contains('This plugin is not configured to enable all '
-                  'Gradle-driven lint warnings and treat them as errors.'),
-              contains('The following packages had errors:'),
-            ],
-          ));
-    });
-
-    test('fails if javac lint-warnings-as-errors is missing', () async {
-      final RepositoryPackage plugin =
-          createFakePlugin('plugin1', packagesDir, extraFiles: <String>[
-        'example/android/gradlew',
-      ], platformSupport: <String, PlatformDetails>{
-        platformAndroid: const PlatformDetails(PlatformSupport.inline)
-      });
-      writeFakePluginBuildGradle(plugin);
-      writeFakeExampleBuildGradle(plugin.getExamples().first, 'plugin1',
-          warningsConfigured: false);
-
-      Error? commandError;
-      final List<String> output = await runCapturingPrint(
-          runner, <String>['lint-android'], errorHandler: (Error e) {
-        commandError = e;
-      });
-
-      expect(commandError, isA<ToolExit>());
-      expect(
-          output,
-          containsAllInOrder(
-            <Matcher>[
-              contains('The example "example" is not configured to treat javac '
-                  'lints and warnings as errors.'),
               contains('The following packages had errors:'),
             ],
           ));