+++ packages/file_selector/file_selector_windows/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 0.9.1+8
 
+* Sets a cmake_policy compatibility version to fix build warnings.
 * Updates minimum Flutter version to 3.3.
 
 ## 0.9.1+7